refactor(results): extract shared DataQualityContent component
Second PR in the live ↔ CRP-2026 shared-components series (after PR
#1866 for top-barriers). Eliminates 100 diff lines of drift between
/results/data-quality and /results/crp-2026/data-quality by introducing
a single <DataQualityContent variant="live" | "crp" data={...} />
component.

Both page.tsx files become 17-line shells that import their tier's
JSON and the shared component. The two pages had drifted in:

- Title and date treatment (LastUpdated vs static "Published" badge)
- Sample list (live shows all 5; CRP filters to the 3 nested samples)
- Sample-count words ("five" vs "three")
- Filter chain cards (live has cards 4 + 5 for All V2 Finished + All
  V2; CRP omits them)
- Constraints prose (long for live, short for CRP)
- Cross-reference link prefixes (/results vs /results/crp-2026)
- Number formatting (CRP wraps in fmt(); live uses ?? '-')

The variant prop drives all of these from one place. The CRP page
inherits the better number formatting (proper integer / decimal
handling) automatically.

Net: +948 / -1650 (net -702 lines).

Tests: 606/606 pass. Build clean. Both routes still prerender as static.
